몬스터 이동 로직 버그 수정
몬스터가 이동시 성벽 밖으로 나가는 버그 수정

+++ Assets/Script/System/EquippedSkillButtonController.cs
@@ -12,7 +12,7 @@ public class EquippedSkillButtonController : MonoBehaviour
     private SkillData  skillData;
     private int skillAttribute;
     private int skillGrade;
-    private int skillCoolTime;
+    private float skillCoolTime;
 
     private bool isOnCooldown = false;
     private string skillName;
